+++ demo/test_dir/web_case/test_po_demo.py
@@ -3,18 +3,18 @@ page object model
 Using the poium Library
 https://github.com/SeldomQA/poium
 ```
-> pip install poium=1.2.0
+> pip install poium>=1.3.2
 ```
 """
 import seldom
-from seldom import Seldom
-from poium import Page, Element
+from poium import Page, Element, Elements
 
 
 class BaiduPage(Page):
     """baidu page"""
-    input = Element(id_="kw", describe="搜索输入框")
-    button = Element(id_="su", describe="搜索按钮")
+    input = Element("#kw", describe="搜索输入框")
+    button = Element("id=su", describe="搜索按钮")
+    result = Elements("//div/h3/a", describe="搜索结果")
 
 
 class BaiduTest(seldom.TestCase):
@@ -24,11 +24,26 @@ class BaiduTest(seldom.TestCase):
         """
         A simple test
         """
-        page = BaiduPage(Seldom.driver, print_log=True)
+        page = BaiduPage()
         page.open("https://www.baidu.com")
+        self.sleep(3)
+
+        # assert element is exist
+        self.assertTrue(page.input.is_exist())
+        self.assertTrue(page.button.is_exist())
+
+        # operation element
         page.input.send_keys("seldom")
         page.button.click()
+        self.sleep(3)
+
+        # assert title
         self.assertTitle("seldom_百度搜索")
+
+        # Loop assertion result
+        for r in page.result:
+            # assert text in
+            self.assertIn("seldom", r.text)
 
 
 if __name__ == '__main__':